fix(some ROS2 pkgs): docker hotfix to permanent
