--- cone_detection_camera/detection/detection/resources/class.names
@@ -1,4 +1,0 @@
-blue
-small_orange
-other
-yellow
