+++ cone_detection_camera/detection/detection/resources/coco.names
@@ -1,0 +1,4 @@
+blue
+small_orange
+other
+yellow
