+++ planner/setup.cfg
@@ -1,0 +1,5 @@
+[develop]
+script_dir=$base/lib/planner
+
+[install]
+install_scripts=$base/lib/planner
